Fix quality flaws.
Change fix done in commit 80470a1cd330e123022091e2de8204b489d49ccf

+++ sonar-php-plugin/src/main/java/org/sonar/plugins/php/phpunit/xml/MetricsNode.java
@@ -39,14 +39,14 @@ public final class MetricsNode {
    */
   @XStreamAsAttribute
   @XStreamAlias("statements")
-  private int totalStatementsCount;
+  private double totalStatementsCount;
 
   /**
    * Gets the coveredstatements.
    *
    * @return the coveredstatements
    */
-  public int getCoveredStatements() {
+  public double getCoveredStatements() {
     return coveredStatements;
   }
 
@@ -55,7 +55,7 @@ public final class MetricsNode {
    *
    * @return the total statements count
    */
-  public int getTotalStatementsCount() {
+  public double getTotalStatementsCount() {
     return totalStatementsCount;
   }
 }
